redo some things
