+++ snapshot/git/gitfiler/checkouter.go
@@ -2,57 +2,10 @@ package gitfiler
 
 import (
 	"fmt"
-	"io/ioutil"
-	"path"
 
-	"github.com/scootdev/scoot/os/temp"
 	"github.com/scootdev/scoot/snapshot"
 	"github.com/scootdev/scoot/snapshot/git/repo"
 )
-
-// Utilities for Reference Repositories.
-// A Reference Repository is a way to clone repos locally so that the clone takes less time and disk space.
-// By passing --reference <local path> to a git clone, the clone will not copy the whole ODB but instead
-// hardlink. This means the clone is much faster and also takes very little extra hard disk space.
-// Cf. https://git-scm.com/docs/git-clone
-
-// RepoGetter lets a client get a Repository to use as a Reference Repository.
-type RepoGetter interface {
-	// Get gets the Repository to use as a reference repository.
-	Get() (*repo.Repository, error)
-}
-
-func NewSingleRepoPool(repoGetter RepoGetter, doneCh chan struct{}) *RepoPool {
-	singlePool := NewRepoPool(nil, nil, doneCh)
-	go func() {
-		r, err := repoGetter.Get()
-		singlePool.Release(r, err)
-	}()
-	return singlePool
-}
-
-func NewSingleRepoCheckouter(repoGetter RepoGetter, doneCh chan struct{}) *Checkouter {
-	pool := NewSingleRepoPool(repoGetter, doneCh)
-	return NewCheckouter(pool)
-}
-
-func NewRefRepoCloningCheckouter(refRepoGetter RepoGetter, clonesDir *temp.TempDir, doneCh chan struct{}) *Checkouter {
-	refPool := NewSingleRepoPool(refRepoGetter, doneCh)
-
-	cloner := &refCloner{refPool: refPool, clonesDir: clonesDir}
-	var clones []*repo.Repository
-	fis, _ := ioutil.ReadDir(clonesDir.Dir)
-	for _, fi := range fis {
-		// TODO(dbentley): maybe we should check that these clones are in fact clones
-		// of the reference repo? Using some kind of git commands to determine its upstream?
-		if clone, err := repo.NewRepository(path.Join(clonesDir.Dir, fi.Name())); err == nil {
-			clones = append(clones, clone)
-		}
-	}
-
-	pool := NewRepoPool(cloner, clones, doneCh)
-	return NewCheckouter(pool)
-}
 
 func NewCheckouter(repos *RepoPool) *Checkouter {
 	return &Checkouter{repos: repos}
@@ -62,6 +15,8 @@ func NewCheckouter(repos *RepoPool) *Checkouter {
 type Checkouter struct {
 	repos *RepoPool
 }
+
+const DEFAULT_REV = "1dda9fbde682e4922a0d5709c5539f573db4cc54"
 
 // Checkout checks out id (a raw git sha) into a Checkout.
 // It does this by making a new clone (via reference) and checking out id.
@@ -79,6 +34,11 @@ func (c *Checkouter) Checkout(id string) (co snapshot.Checkout, err error) {
 			c.repos.Release(repo, repoErr)
 		}
 	}()
+
+	// TODO(dbentley): do more ot validate id. E.g., don't let "HEAD" or "master" slip through
+	if id == "" {
+		id = DEFAULT_REV
+	}
 
 	// -d removes directories. -x ignores gitignore and removes everything.
 	// -f is force. -f the second time removes directories even if they're git repos themselves
